chore(playwright): externalize common functions and constants in dedicated utils file
