+++ playwright/e2e-tests/utils.ts
@@ -1,0 +1,39 @@
+import type { Page } from '@playwright/test'
+import { expect } from '@playwright/test'
+
+// Retrieve frontend URL from environment variables (see playwright.config.ts)
+export const clientURL = process.env.KEYCLOAK_REDIRECT_URI || 'http://change-me'
+
+interface Credentials {
+  username: string
+  password: string
+}
+
+// Users referenced in Keycloak dev realm (../keycloak/realms/realm-dev.json)
+export const adminUser: Credentials = {
+  username: 'admin',
+  password: 'admin',
+}
+export const testUser: Credentials = {
+  username: 'test',
+  password: 'test',
+}
+export const cnolletUser: Credentials = {
+  username: 'cnollet',
+  password: 'test',
+}
+
+export async function signInCloudPiNative({
+  page,
+  credentials,
+}: {
+  page: Page
+  credentials: Credentials
+}) {
+  const { username, password } = credentials
+  await page.getByRole('link', { name: 'Se connecter' }).click()
+  await page.locator('#username').fill(username)
+  await page.locator('#password').fill(password)
+  await page.locator('#kc-login').click()
+  await expect(page.locator('#top')).toContainText('Cloud π Native')
+}
